page de profil

+++ delete-article.php
@@ -1,9 +1,17 @@
 <?php
-require_once __DIR__.'/database/database.php';
-$articleDB = require_once __DIR__ . '/database/models/articleDB.php';
-$_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-$id = $_GET['id'] ?? '';
-if ($id) {
-    $articleDB->deleteOne($id);
+require_once __DIR__ . '/database/database.php';
+require_once __DIR__ . '/database/security.php';
+$currentUser = isLoggedin();
+if ($currentUser) {
+    $articleDB = require_once __DIR__ . '/database/models/ArticleDB.php';
+    $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+    $id = $_GET['id'] ?? '';
+    if ($id) {
+        $article = $articleDB->fetchOne($id);
+        if ($article['author'] === $currentUser['id']) {
+            $articleDB->deleteOne($id);
+        }
+    }
 }
+
 header('Location: /');
